EogScrollView: add image rotation gesture
GtkGestureRotate is used to determine angle changes, which are then
used to rotate the image in 90 degree steps. In order to avoid
unintended angle changes when the delta is too close to a switch
point, there is a "grey area" threshold that makes it harder to switch
forth and back again.

This gesture has been added to the same group than the GtkGestureZoom,
as both are meant to share sequences and state.